// src/interfaces/motion.interface.ts
export interface transitionProps {
    delay?: number;
    duration?: number;
    ease?: string;
    type?: string;
    damping?: number;
    mass?: number;
    stiffness?: number;
}
export interface transformProps {
    offsetX?: string;
    offsetY?: string;
    targetX?: string;
    targetY?: string;
}